playwright_test/pyproject.toml
[tool.poetry]
name = "playwright-test"
version = "0.1.0"
description = ""
authors = ["zubby--samsung <zubbyik@gmail.com>"]
readme = "README.md"
packages = [{include = "playwright_test"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.3.0"
allure-pytest = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
